XWIKI-10431: In major installations, Filesystem Attachment Store deleteEmptyDirs() can block startup for significant time. Allow disabling of deleteEmptyDirs()

+++ xwiki-platform-core/xwiki-platform-store/xwiki-platform-store-filesystem-attachments/src/main/java/org/xwiki/store/filesystem/internal/DefaultFilesystemStoreTools.java
@@ -102,6 +102,9 @@ public class DefaultFilesystemStoreTools implements FilesystemStoreTools, Initia
     @Named("path")
     private EntityReferenceSerializer<String> pathSerializer;
 
+    @Inject
+    private FilesystemAttachmentsConfiguration config;
+
     /**
      * We need to get the XWiki object in order to get the work directory.
      */
@@ -156,7 +159,9 @@ public class DefaultFilesystemStoreTools implements FilesystemStoreTools, Initia
     public void initialize()
     {
         this.storageDir = new File(this.environment.getPermanentDirectory(), STORAGE_DIR_NAME);
-        deleteEmptyDirs(this.storageDir, 0);
+        if (config.cleanOnStartup()) {
+            deleteEmptyDirs(this.storageDir, 0);
+        }
     }
 
     /**
